feat: API to assign story owners
When using the query builder to insert story, the function .values() only accept the instance of the Story entity class.

If u use .values( {title, type, points, description, project: { project_id: projectID }}), it will throw the error "story_id is null". This is because the passed story object does not have the beforeInsert() function